feat(ratelimit): move rate limiting logic to middleware
