+++ apps/api/src/transactions/transaction.module.ts
@@ -1,11 +1,16 @@
-import { Module } from '@nestjs/common';
+import { Module, MiddlewareConsumer, NestModule } from '@nestjs/common';
 
 import { TransactionsService } from './transaction.service';
 import { TransactionsController } from './transaction.controller';
+import { RateLimitMiddleware } from './rate-limit.middleware';
 
 @Module({
   controllers: [TransactionsController],
   providers: [TransactionsService],
   exports: [TransactionsService],
 })
-export class TransactionsModule {}
+export class TransactionsModule implements NestModule {
+  configure(consumer: MiddlewareConsumer) {
+    consumer.apply(RateLimitMiddleware).forRoutes(TransactionsController);
+  }
+}
