+++ apps/api/src/transactions/transaction.service.ts
@@ -8,46 +8,15 @@ import { mockData } from './transactions.mock';
 import { CreateTransactionDto } from '@repo/api/transactions/dto/create-transaction.dto';
 import { UpdateTransactionDto } from '@repo/api/transactions/dto/update-transaction.dto';
 
-type RateLimiter = {
-  canMakeRequest(): boolean;
-  recordRequest(): void;
-};
-
 // TODO: A lot of these actions should be internally wrapped in try/catch blocks
 //       I haven't done so here because we're not interfacing with an actual database
 
 @Injectable()
-export class TransactionsService implements RateLimiter {
+export class TransactionsService {
   private readonly _transactions: Transaction[] = mockData;
-  private readonly RATE_LIMIT = 5; // 5 requests per minute
-  private readonly RATE_WINDOW = 60_000; // 1 minute in ms
   private readonly BATCH_SIZE = 1_000;
-  private requestCount = 0;
-  private requestWindowStart = Date.now();
-
-  // TODO: The rate limiting should be handled through middleware, but I am not familiar
-  //       enough wih Nestjs to have figured that out yet, nor do I have enough time
-  canMakeRequest() {
-    const now = Date.now();
-    if (now - this.requestWindowStart > this.RATE_WINDOW) {
-      this.requestCount = 0;
-      this.requestWindowStart = now;
-    }
-
-    return this.requestCount < this.RATE_LIMIT;
-  }
-
-  recordRequest() {
-    this.requestCount++;
-  }
 
   create(createTransactionDto: CreateTransactionDto) {
-    if (!this.canMakeRequest()) {
-      throw new Error('Rate limit exceeded');
-    } else {
-      this.recordRequest();
-    }
-
     const newTransaction = {
       ...createTransactionDto,
       createdAt: new Date().toISOString(),
@@ -62,12 +31,6 @@ export class TransactionsService implements RateLimiter {
   // Everything is typed as strings here because of how query params work
   // They could technically also be arrays but let's not worry about that now
   findAll(page?: string, limit?: string, startDate?: string, endDate?: string) {
-    if (!this.canMakeRequest()) {
-      throw new Error('Rate limit exceeded');
-    } else {
-      this.recordRequest();
-    }
-
     const itemsPerPage = parseInt(limit ?? String(this.BATCH_SIZE), 10);
     const currentPage = parseInt(page ?? '1', 10);
     let filteredTransactions = this._transactions;
@@ -109,12 +72,6 @@ export class TransactionsService implements RateLimiter {
   }
 
   findOne(id: string) {
-    if (!this.canMakeRequest()) {
-      throw new Error('Rate limit exceeded');
-    } else {
-      this.recordRequest();
-    }
-
     const found = this._transactions.find(({ id: _id }) => _id === id);
 
     if (!found) {
@@ -125,34 +82,16 @@ export class TransactionsService implements RateLimiter {
   }
 
   findByUserId(userId: string) {
-    if (!this.canMakeRequest()) {
-      throw new Error('Rate limit exceeded');
-    } else {
-      this.recordRequest();
-    }
-
     return this._transactions.filter(
       ({ userId: _userId }) => _userId === userId,
     );
   }
 
   findByType(type: 'payout' | 'spent' | 'earned') {
-    if (!this.canMakeRequest()) {
-      throw new Error('Rate limit exceeded');
-    } else {
-      this.recordRequest();
-    }
-
     return this._transactions.filter(({ type: _type }) => _type === type);
   }
 
   update(id: string, updateTransactionDto: UpdateTransactionDto) {
-    if (!this.canMakeRequest()) {
-      throw new Error('Rate limit exceeded');
-    } else {
-      this.recordRequest();
-    }
-
     const index = this._transactions.findIndex(({ id: _id }) => _id === id);
 
     if (index === -1) {
@@ -171,12 +110,6 @@ export class TransactionsService implements RateLimiter {
   }
 
   remove(id: string) {
-    if (!this.canMakeRequest()) {
-      throw new Error('Rate limit exceeded');
-    } else {
-      this.recordRequest();
-    }
-
     const index = this._transactions.findIndex(({ id: _id }) => _id === id);
 
     if (index === -1) {
